+++ cranelift/codegen/src/isa/aarch64/lower/isle.rs
@@ -10,8 +10,8 @@ use super::{
     fp_reg, lower_condcode, lower_fp_condcode, stack_reg, writable_link_reg, writable_zero_reg,
     zero_reg, ASIMDFPModImm, ASIMDMovModImm, BranchTarget, CallIndInfo, CallInfo, Cond, CondBrKind,
     ExtendOp, FPUOpRI, FPUOpRIMod, FloatCC, Imm12, ImmLogic, ImmShift, Inst as MInst, IntCC,
-    JTSequenceInfo, MachLabel, MemLabel, MoveWideConst, MoveWideOp, Opcode, OperandSize, Reg,
-    SImm9, ScalarSize, ShiftOpAndAmt, UImm12Scaled, UImm5, VecMisc2, VectorSize, NZCV,
+    MachLabel, MemLabel, MoveWideConst, MoveWideOp, Opcode, OperandSize, Reg, SImm9, ScalarSize,
+    ShiftOpAndAmt, UImm12Scaled, UImm5, VecMisc2, VectorSize, NZCV,
 };
 use crate::ir::condcodes;
 use crate::isa;
@@ -43,7 +43,6 @@ type BoxCallInfo = Box<CallInfo>;
 type BoxCallIndInfo = Box<CallIndInfo>;
 type BoxReturnCallInfo = Box<ReturnCallInfo>;
 type VecMachLabel = Vec<MachLabel>;
-type BoxJTSequenceInfo = Box<JTSequenceInfo>;
 type BoxExternalName = Box<ExternalName>;
 type VecArgPair = Vec<ArgPair>;
 
@@ -596,32 +595,15 @@ impl Context for IsleContext<'_, '_, MInst, AArch64Backend> {
         super::regs::pinned_reg().to_real_reg().unwrap().into()
     }
 
-    fn branch_target(&mut self, elements: &VecMachLabel, idx: u8) -> BranchTarget {
-        BranchTarget::Label(elements[idx as usize])
+    fn branch_target(&mut self, label: MachLabel) -> BranchTarget {
+        BranchTarget::Label(label)
     }
 
-    fn targets_jt_size(&mut self, elements: &VecMachLabel) -> u32 {
-        (elements.len() - 1) as u32
-    }
-
-    fn targets_jt_space(&mut self, elements: &VecMachLabel) -> CodeOffset {
+    fn targets_jt_space(&mut self, elements: &BoxVecMachLabel) -> CodeOffset {
         // calculate the number of bytes needed for the jumptable sequence:
         // 4 bytes per instruction, with 8 instructions base + the size of
         // the jumptable more.
-        4 * (8 + self.targets_jt_size(elements))
-    }
-
-    fn targets_jt_info(&mut self, elements: &VecMachLabel) -> BoxJTSequenceInfo {
-        let targets: Vec<BranchTarget> = elements
-            .iter()
-            .skip(1)
-            .map(|bix| BranchTarget::Label(*bix))
-            .collect();
-        let default_target = BranchTarget::Label(elements[0]);
-        Box::new(JTSequenceInfo {
-            targets,
-            default_target,
-        })
+        (4 * (8 + elements.len())).try_into().unwrap()
     }
 
     fn min_fp_value(&mut self, signed: bool, in_bits: u8, out_bits: u8) -> Reg {
