+++ cranelift/codegen/src/isa/riscv64/inst.isle
@@ -2625,7 +2625,7 @@
 (decl int_zero_reg (Type) ValueRegs)
 (extern constructor int_zero_reg int_zero_reg)
 
-(decl lower_cond_br (IntCC ValueRegs VecMachLabel Type) Unit)
+(decl lower_cond_br (IntCC ValueRegs MachLabelSlice Type) Unit)
 (extern constructor lower_cond_br lower_cond_br)
 
 ;; Convert a truthy value, possibly of more than one register (an I128), to
@@ -2644,13 +2644,11 @@
 
 (decl label_to_br_target (MachLabel) CondBrTarget)
 (extern constructor label_to_br_target label_to_br_target)
+(convert MachLabel CondBrTarget label_to_br_target)
 
-(decl vec_label_get (VecMachLabel u8) MachLabel)
-(extern constructor vec_label_get vec_label_get)
-
-(decl partial lower_branch (Inst VecMachLabel) Unit)
-(rule (lower_branch (jump _) targets )
-      (emit_side_effect (SideEffectNoResult.Inst (MInst.Jal (vec_label_get targets 0)))))
+(decl partial lower_branch (Inst MachLabelSlice) Unit)
+(rule (lower_branch (jump _) (single_target label))
+      (emit_side_effect (SideEffectNoResult.Inst (MInst.Jal label))))
 
 ;; Default behavior for branching based on an input value.
 (rule (lower_branch (brif v @ (value_type (fits_in_64 ty)) _ _) targets)
@@ -2660,21 +2658,17 @@
 
 ;; Branching on the result of an fcmp
 (rule 1
-  (lower_branch (brif (maybe_uextend (fcmp cc a @ (value_type ty) b)) _ _) targets)
+  (lower_branch (brif (maybe_uextend (fcmp cc a @ (value_type ty) b)) _ _) (two_targets then else))
   (if-let $true (floatcc_unordered cc))
-  (let ((then CondBrTarget (label_to_br_target (vec_label_get targets 0)))
-        (else CondBrTarget (label_to_br_target (vec_label_get targets 1))))
-    (emit_side_effect (cond_br (emit_fcmp (floatcc_complement cc) ty a b) else then))))
+  (emit_side_effect (cond_br (emit_fcmp (floatcc_complement cc) ty a b) else then)))
 
 (rule 1
-  (lower_branch (brif (maybe_uextend (fcmp cc a @ (value_type ty) b)) _ _) targets)
+  (lower_branch (brif (maybe_uextend (fcmp cc a @ (value_type ty) b)) _ _) (two_targets then else))
   (if-let $false (floatcc_unordered cc))
-  (let ((then CondBrTarget (label_to_br_target (vec_label_get targets 0)))
-        (else CondBrTarget (label_to_br_target (vec_label_get targets 1))))
-    (emit_side_effect (cond_br (emit_fcmp cc ty a b) then else))))
+  (emit_side_effect (cond_br (emit_fcmp cc ty a b) then else)))
 
 
-(decl lower_br_table (Reg VecMachLabel) Unit)
+(decl lower_br_table (Reg MachLabelSlice) Unit)
 (extern constructor lower_br_table lower_br_table)
 
 (rule (lower_branch (br_table index _) targets)
